tests/requirements.txt
django_nose
django-bootstrap-form
https://github.com/InfoAgeTech/django-core/tarball/master
https://github.com/InfoAgeTech/django-generics/tarball/master
https://github.com/InfoAgeTech/django-testing/tarball/master